+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- `mark_left` function to mark removed members as `left`.
+
 ## [2.4.5] - 2024-06-24
 
 ### Fixed
@@ -13,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Invalid events parsing.
 
 ## [2.4.4] - 2024-04-09
+
 ### Fixed
 
 - Invalid payload parsing in anti entropy step.